pr-compliance-guard/compliance/config.py
#import yaml
#
#def load_config(path="compliance.yaml"):
#    with open(path, "r") as f:
#        return yaml.safe_load(f)
#

#pr-compliance-guard/compliance/config.py

import yaml


def load_config(path: str) -> dict:
    with open(path, "r") as f:
        return yaml.safe_load(f)

